Refactor Excel builder into dedicated tool package
